+++ src/main/kotlin/com/github/victorrentea/livecoding/assertj/MigrateToAssertJInspection.kt
@@ -16,24 +16,6 @@ class MigrateToAssertJInspection : BaseInspection() {
         const val FIX_NAME = "Replace with 'org.assertj.core.api.Assertions' method call (AssertJ)"
         private val log = logger<MigrateToAssertJInspection>()
 
-
-        val SIGNATURES = mapOf<String, (PsiMethodCallExpression) -> String>(
-            "assertEquals(Object,Object)" to ::migrateEquals,
-            "assertEquals(int,Integer)" to ::migrateEquals,
-            "assertEquals(int,int)" to ::migrateEquals,
-            "assertEquals(float,float)" to ::migrateEquals,
-            "assertEquals(double,double)" to ::migrateEquals,
-            "assertEquals(char,char)" to ::migrateEquals,
-            "assertEquals(char,char)" to ::migrateEquals,
-            "assertEquals(long,long)" to ::migrateEquals,
-            "assertEquals(long,Long)" to ::migrateEquals,
-            "assertFalse(boolean)" to ::migrateFalse,
-            "assertTrue(boolean)" to ::migrateTrue,
-            "assertNull(Object)" to ::migrateNull,
-            "assertNotNull(Object)" to ::migrateNotNull,
-            "assertThrows(Class<T>,Executable)" to ::migrateThrows,
-        )
-
         private fun tryImportStatically(fqMethodName: String, context: PsiElement): String {
             val fqClassName = fqMethodName.substringBeforeLast(".")
             val methodName = fqMethodName.substringAfterLast(".")
@@ -45,42 +27,14 @@ class MigrateToAssertJInspection : BaseInspection() {
             return methodPrefix + methodName
         }
 
-        private fun migrateEquals(call: PsiMethodCallExpression): String {
-            return "org.assertj.core.api.Assertions.assertThat(" + call.argumentList.expressions[1].text +
-                    ").isEqualTo(" + call.argumentList.expressions[0].text + ")"
-        }
-
-        private fun migrateNull(call: PsiMethodCallExpression): String {
-            return "org.assertj.core.api.Assertions.assertThat(" + call.argumentList.expressions[0].text + ").isNull()"
-        }
-
-        private fun migrateFalse(call: PsiMethodCallExpression): String {
-            return "org.assertj.core.api.Assertions.assertThat(" + call.argumentList.expressions[0].text + ").isFalse()"
-        }
-
-        private fun migrateTrue(call: PsiMethodCallExpression): String {
-            return "org.assertj.core.api.Assertions.assertThat(" + call.argumentList.expressions[0].text + ").isTrue()"
-        }
-
-        private fun migrateNotNull(call: PsiMethodCallExpression): String {
-            return "org.assertj.core.api.Assertions.assertThat(" + call.argumentList.expressions[0].text + ").isNotNull()"
-        }
-
-        private fun migrateThrows(call: PsiMethodCallExpression): String {
-            return "org.assertj.core.api.Assertions.assertThatThrownBy(" + call.argumentList.expressions[1].text + ")\n" +
-                    ".isInstanceOf(" + call.argumentList.expressions[0].text + ")"
-        }
     }
 
-    override fun shouldInspect(file: PsiFile?): Boolean {
-        return file?.let { FrameworkDetector.assertjIsPresent(it) } == true
-    }
+    override fun shouldInspect(file: PsiFile?) =
+        file?.let { FrameworkDetector.assertjIsPresent(it) } == true
 
     override fun buildErrorString(vararg infos: Any?) = INSPECTION_NAME
 
-    override fun buildVisitor(): BaseInspectionVisitor {
-        return MigrateToAssertJVisitor()
-    }
+    override fun buildVisitor() = MigrateToAssertJVisitor()
 
     class MigrateToAssertJVisitor : BaseInspectionVisitor() {
         override fun visitMethodCallExpression(expression: PsiMethodCallExpression?) {
@@ -88,26 +42,36 @@ class MigrateToAssertJInspection : BaseInspection() {
 
             if (!calledPsiMethod.hasModifierProperty(PsiModifier.STATIC)) return
             val classQName = calledPsiMethod.containingClass?.qualifiedName ?: return
-            if (classQName != "org.junit.jupiter.api.Assertions" && classQName != "org.junit.Assert") return
 
-            val typeStr = calledPsiMethod.parameterList.parameters
-                .map { it.typeElement?.type?.presentableText }.joinToString(",")
+            if (classQName !in listOf("org.junit.jupiter.api.Assertions", "org.junit.Assert")) return
 
-            val methodSignature = calledPsiMethod.name + "(" + typeStr + ")"
+            val descr = if (calledPsiMethod.parameterList.parameters.last()
+                    .let{it.name == "message" && it.type.equalsToText("java.lang.String")}) {
+                "\n.describedAs(" + expression.argumentList.expressions.last().text+")\n"
+            } else ""
 
-            log.debug("Found $methodSignature")
-
-            if (SIGNATURES.containsKey(methodSignature)) {
-                registerError(expression.methodExpression, methodSignature)
+            val args = expression.argumentList.expressions
+            val arg0 = args[0].text
+            val arg1 = args.getOrNull(1)?.text
+            val migratedCode = "org.assertj.core.api.Assertions." + when (calledPsiMethod.name)  {
+                "assertEquals" -> "assertThat($arg1)$descr.isEqualTo($arg0)"
+                "assertNotEquals" -> "assertThat($arg1)$descr.isNotEqualTo($arg0)"
+                "assertTrue" -> "assertThat($arg0)$descr.isTrue()"
+                "assertFalse" -> "assertThat($arg0)$descr.isFalse()"
+                "assertNull" -> "assertThat($arg0)$descr.isNull()"
+                "assertNotNull" -> "assertThat($arg0)$descr.isNotNull()"
+                "assertThrows" -> if (descr == "") "assertThatThrownBy($arg1).isInstanceOf($arg0)"
+                                    else "assertThatThrownBy($arg1, "+expression.argumentList.expressions.last().text+").isInstanceOf($arg0)"
+                else -> return
             }
+
+            registerError(expression.methodExpression, migratedCode)
         }
     }
 
-    override fun buildFix(vararg infos: Any?): InspectionGadgetsFix {
-        return MigrateToAssertJFix(infos[0] as String)
-    }
+    override fun buildFix(vararg infos: Any?) = MigrateToAssertJFix(infos[0] as String)
 
-    class MigrateToAssertJFix(private val signature: String) : InspectionGadgetsFix() {
+    class MigrateToAssertJFix(private val rawMigratedCode: String) : InspectionGadgetsFix() {
         override fun getFamilyName(): String {
             return FIX_NAME
         }
@@ -115,29 +79,15 @@ class MigrateToAssertJInspection : BaseInspection() {
         override fun doFix(project: Project?, descriptor: ProblemDescriptor?) {
             if (project == null) return
             val methodCall = descriptor?.psiElement?.parent as? PsiMethodCallExpression ?: return
-            val documentManager = PsiDocumentManager.getInstance(project)
-            val document = documentManager.getDocument(descriptor.psiElement.containingFile) ?: return
-            val transformFunction = SIGNATURES[signature] ?: return
-//            WriteCommandAction.runWriteCommandAction(project,
-//                FIX_NAME, "Live-Coding", {
-            val rawExpressionText = transformFunction(methodCall)
 
-            val fqMethodPart = rawExpressionText.substringBefore("(")
+            val fqMethodPart = rawMigratedCode.substringBefore("(")
             val shortMethod = tryImportStatically(fqMethodPart, methodCall)
-            val newExpressionText = shortMethod + "(" + rawExpressionText.substringAfter("(")
+            val newExpressionText = shortMethod + "(" + rawMigratedCode.substringAfter("(")
 
             val newExpression =
                 JavaPsiFacade.getElementFactory(project).createExpressionFromText(newExpressionText, methodCall)
 
             methodCall.replace(newExpression)
-//                    documentManager.doPostponedOperationsAndUnblockDocument(document)
-//
-//                    JavaCodeStyleManager.getInstance(project).optimizeImports(newExpression.containingFile)
-//
-//                    documentManager.doPostponedOperationsAndUnblockDocument(document)
-//
-//                    JavaCodeStyleManager.getInstance(project).shortenClassReferences(newExpression)
-//                })
         }
 
     }