+++ src/test/live-coding-playground/src/main/java/assertj/migrate/AssertEqualsInt.java
@@ -1,12 +1,13 @@
-package assertj;
+package assertj.migrate;
 
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
 
-public class AssertEqualsLong {
+public class AssertEqualsInt {
     @Test
     public void test() {
-        long actual = 1L;
+        int actual = 1;
         Assertions.assertEquals(1, actual); //
+        Assertions.assertEquals(1, actual, "Message"); //
     }
 }