+++ src/test/live-coding-playground/src/main/java/assertj/migrate/AssertFalse.java
@@ -1,4 +1,4 @@
-package assertj;
+package assertj.migrate;
 
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;