+++ src/test/live-coding-playground/src/main/java/assertj/migrate/AssertThrows.java
@@ -1,12 +1,15 @@
-package assertj;
+package assertj.migrate;
 
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
+
+import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 public class AssertThrows {
     @Test
     public void test() {
         Assertions.assertThrows(IllegalArgumentException.class, () -> f()); //
+        Assertions.assertThrows(IllegalArgumentException.class, () -> f(), "text");
     }
 
     void f() {