+++ src/test/live-coding-playground/src/main/java/assertj/migrate/AssertTrue.java
@@ -1,7 +1,9 @@
-package assertj;
+package assertj.migrate;
 
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
+
+import static org.assertj.core.api.Assertions.assertThat;
 
 public class AssertTrue {
     @Test