+++ CHANGELOG.md
@@ -87,3 +87,7 @@ a first stable version
 - remove 'color test files UX' - proved not really helpful.
 - Added background mood to ask more people to open their webcam (Ctrl-alt-6)
 - Inspection to replace usages of JUnit5 Assertions or JUnit4 Assert with AssertJ Assertions, aka. assertThat(actual).isEqualTo(expected); - correct, but not exhaustive
+
+## 1.0.1 
+- Add Optimize AssertJ assertions inspection 
+- Improve Junit->Assertions migration inspection (more api included): added description and isNotEqualTo